+++ packages/puer_time_travel/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 - Export `puer` from this package (#24)
+- Assert on unique names in time travel controller (#39)
 
 ## 1.0.0-alpha.2